% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rparquet-package.R
\docType{package}
\name{rparquet-package}
\alias{rparquet}
\alias{rparquet-package}
\title{rparquet: Read and Write 'Parquet' Files}
\description{
Reads and writes 'Parquet' files.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
