% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RParquet.R
\docType{package}
\name{RParquet}
\alias{RParquet}
\alias{RParquet-package}
\title{RParquet: A package for read and write parquet file in R}
\description{
The RParquet package provides two basic functions:
rparquet_reader, rparquet_writer
}
