% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RParquet.R
\name{rparquet_metadata}
\alias{rparquet_metadata}
\title{Generate the parquet metadata information}
\usage{
rparquet_metadata(filename, details = F)
}
\arguments{
\item{filename}{- A string. Specifies the name of the parquet file}

\item{details}{- detailed information about the parquet file}
}
\value{
DataFrame
}
\description{
This function returns a metadata summary based on the input parquet file.
}
\examples{
\dontrun{
f <- "path_to_file.parquet"
df <- rparquet_metadata(f)

}
}
