% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RParquet.R
\name{rparquet_reader}
\alias{rparquet_reader}
\title{Read the R DataFrame from a parquet file}
\usage{
rparquet_reader(filename, columns = c(-1), filter = c(TRUE),
  row_size = 1e+05, threads = 0, verbose = 0)
}
\arguments{
\item{filename}{- A string. Specifies the name of the parquet file}

\item{columns}{- An integer vector. Specifies the wanted columns. default is to select all the columns}

\item{filter}{- A logical vector. Specifies T/F for each row. default is all row selected}

\item{row_size}{- An integer. Specify num of rows to read per each read action.}

\item{threads}{- An integer. Specify read columns with the indicated level of parallelism.}

\item{verbose}{- An integer. 0-no verbose output, 1-regular verbose output, including
row/col/type etc}
}
\value{
DataFrame
}
\description{
This function returns a dataframe based on the columns selected in the parquet file.
It supports "INT32", "INT64", "TIMESTAMP", "DOUBLE", "STRING", "BOOL" Apache Arrow types.
}
\examples{
\dontrun{
f <- "path_to_file.parquet"

df <- rparquet_reader(f)

cols <- c(1,2,4,7,8,10)
df <- rparquet_reader(filename, columns)

fdf <- read.csv("parth_to_file.csv");
filter = (fdf$sid > 50031571 & fdf$sid != 2018)
df <- rparquet_reader(filename, columns, filter)
}
}
