% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RParquet.R
\name{rparquet_writer}
\alias{rparquet_writer}
\title{Write the R DataFrame into a parquet file}
\usage{
rparquet_writer(df, filename, columns = c(-1), group_rows = 1e+06,
  verbose = 0)
}
\arguments{
\item{df}{- A DataFrame to write}

\item{filename}{- A string. Specifies the parquet filename}

\item{columns}{- A integer vector. Specifies the wanted columns, first colum is 1.
Default value is to select all the columns.}

\item{group_rows}{- A integer. Specifies num of rows per group.}

\item{verbose}{- A integer.0-no verbose output, 1-regular verbose output, including
row/col/type etc}
}
\value{
0
}
\description{
This function writes the dataframe based on the columns selected to the parquet file.
It supports "integer", "integer64", "nanotime", "numeric", "character", "logical" types.
}
\examples{
\dontrun{
f <- "path_to_file.csv"
df <- read.csv(f)
rparquet_writer(df, f)

rparquet_writer(df, f, columns = c(1,2), group_rows = 100)
}
}
